requirements.txt
discord.py==1.7.3
just_config==20.3
google-cloud-datastore

pip~=20.3.4
wheel~=0.34.2
cryptography~=3.3.2
keyring~=22.2.0
attrs~=21.2.0
idna~=2.10
multidict~=5.1.0
MarkupSafe~=1.1.1
click~=7.1.2
blinker~=1.4
yarl~=1.6.3
aiohttp~=3.7.4.post0
chardet~=4.0.0
discord~=1.7.3
setuptools~=52.0.0
future~=0.18.2